+++ python/mujoco/render.cc
@@ -54,9 +54,6 @@ class MjWrapper<raw::MjrContext> : public WrapperBase<raw::MjrContext> {
   X(auxFBO_r);
   X(auxColor);
   X(auxColor_r);
-  X(mat_texid);
-  X(mat_texuniform);
-  X(mat_texrepeat);
   X(textureType);
   X(texture);
   X(skinvertVBO);
@@ -98,9 +95,6 @@ MjrContextWrapper::MjWrapper()
       X(auxFBO_r),
       X(auxColor),
       X(auxColor_r),
-      X(mat_texid),
-      X(mat_texuniform),
-      X(mat_texrepeat),
       X(textureType),
       X(texture),
       X_SKIN(skinvertVBO),
@@ -125,9 +119,6 @@ MjrContextWrapper::MjWrapper(const MjModelWrapper& model, int fontscale)
       X(auxFBO_r),
       X(auxColor),
       X(auxColor_r),
-      X(mat_texid),
-      X(mat_texuniform),
-      X(mat_texrepeat),
       X(textureType),
       X(texture),
       X_SKIN(skinvertVBO),
@@ -228,9 +219,6 @@ PYBIND11_MODULE(_render, pymodule, pybind11::mod_gil_not_used()) {
   X(auxFBO_r);
   X(auxColor);
   X(auxColor_r);
-  X(mat_texid);
-  X(mat_texuniform);
-  X(mat_texrepeat);
   X(textureType);
   X(texture);
   X(skinvertVBO);