+++ src/render/classic/render_context.c
@@ -14,7 +14,6 @@
 
 #include "render/classic/render_context.h"
 
-#include <inttypes.h>  // IWYU pragma: keep
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -1299,44 +1298,6 @@ static void makeFont(mjrContext* con, int fontscale) {
   }
 }
 
-// make materials, just for those that have textures
-static void makeMaterial(const mjModel* m, mjrContext* con) {
-  memset(con->mat_texid, -1, sizeof(con->mat_texid));
-  memset(con->mat_texuniform, 0, sizeof(con->mat_texuniform));
-  memset(con->mat_texrepeat, 0, sizeof(con->mat_texrepeat));
-
-  // find skybox texture
-  for (int i=0; i < m->ntex; i++) {
-    if (m->tex_type[i] == mjTEXTURE_SKYBOX) {
-      if (m->nmat >= mjMAXMATERIAL-2) {
-        mju_error("With skybox, maximum number of materials is %d, got %" PRId64,
-                  mjMAXMATERIAL-1, m->nmat);
-      }
-      for (int j=0; j < mjNTEXROLE; j++) {
-        con->mat_texid[mjNTEXROLE * (mjMAXMATERIAL-1) + j] = -1;
-      }
-      con->mat_texid[mjNTEXROLE * (mjMAXMATERIAL-1) + mjTEXROLE_RGB] = i;
-
-      break;
-    }
-  }
-
-  if (m->nmat >= mjMAXMATERIAL-1) {
-    mju_error("Maximum number of materials is %d, got %" PRId64, mjMAXMATERIAL, m->nmat);
-  }
-  for (int i=0; i < m->nmat; i++) {
-    if (m->mat_texid[i*mjNTEXROLE + mjTEXROLE_RGB] >= 0) {
-      for (int j=0; j < mjNTEXROLE; j++) {
-        con->mat_texid[i*mjNTEXROLE + j] = m->mat_texid[i*mjNTEXROLE + j];
-      }
-      con->mat_texuniform[i] = m->mat_texuniform[i];
-      con->mat_texrepeat[2*i] = m->mat_texrepeat[2*i];
-      con->mat_texrepeat[2*i+1] = m->mat_texrepeat[2*i+1];
-    }
-  }
-}
-
-
 // make textures
 static void makeTexture(const mjModel* m, mjrContext* con) {
   // checks size
@@ -1639,7 +1600,6 @@ void mjr_makeContext_offSize(const mjModel* m, mjrContext* con, int fontscale,
   // make everything
   makeOff(con);
   makeShadow(m, con);
-  makeMaterial(m, con);
   makeTexture(m, con);
   makePlane(m, con);
   makeMesh(m, con);